Rename Icon.js to Icons.js
This better conveys the named exports in Icons (there is no "Icon"
export) and makes room for a new Icon component
